backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
feedparser==6.0.10
